Configure routing mailbox and integrate with StartApp
This introduces a new mailbox in the router, `routerMailbox`, which
serves as a place for the `linkTo` function to send paths to navigate
to.

We have to use a custom `onClick'` handler for these as we don't
want normal browser behavior (e.g. traditional URL navigation) to occur.
Instead, we prevent default behavior and deliver the message
(`path : String`) to the `routerMailbox.address`.

Finally, within `Main.elm`, we configure a second input for the
StartApp, `navigations : Signal Action`, which maps over the router
mailbox and converts strings to the appropriate signal of `NavigateTo
path`.